+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Add info message for SSPX parts when the IVAs are not installed
 * Add support for ReStockPlus spherical pod parts when Making History is not installed
 * Fix incorrect part name for orbital engine cluster from Near Future Spacecraft
+* Fix issues with B9PSConditionalProp when applied to hatches.  Props are now deleted instead of deactivated if their subtype is not active.
 
 
 ## 0.2.20.2 - 2025-10-24